Rename test and refactor
