+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "nodemon --exec babel-node src/index.js",
     "start": "node build/index.js",
-    "test": "mocha --require @babel/register ./test --recursive --timeout 15000 --exit",
+    "test": "mocha --require @babel/register ./test --recursive --timeout 150000 --exit",
     "build": "rm -rf ./build && babel -d ./build ./src -s",
     "migrate": "node src/database/migrate createAllTables",
     "migrate-refresh": "node src/database/migrate dropAllTables",
